+++ src/Conreign.Core.Contracts/Gameplay/Data/MovingFleetData.cs
@@ -5,7 +5,7 @@ namespace Conreign.Core.Contracts.Gameplay.Data
     public class MovingFleetData
     {
         public FleetData Fleet { get; set; } = new FleetData();
-        public long Position { get; set; }
-        public List<long> Route { get; set; } = new List<long>();
+        public int Position { get; set; }
+        public List<int> Route { get; set; } = new List<int>();
     }
 }
