+++ src/Conreign.Core/AI/Battle/ReadOnlyMap.cs
@@ -15,14 +15,14 @@ namespace Conreign.Core.AI.Battle
             _map = map ?? throw new ArgumentNullException(nameof(map));
         }
 
-        public ReadOnlyPlanetData this[long position] => 
+        public ReadOnlyPlanetData this[int position] => 
             new ReadOnlyPlanetData(_map[position], position);
 
         public ReadOnlyPlanetData this[int x, int y] => 
             new ReadOnlyPlanetData(_map[x, y], new Coordinate(x, y, Width, Height).Position);
 
         public int Height => _map.Height;
-        public long MaxDistance => _map.MaxDistance;
+        public int MaxDistance => _map.MaxDistance;
         public int Width => _map.Width;
 
         public IEnumerator<ReadOnlyPlanetData> GetEnumerator()
@@ -35,7 +35,7 @@ namespace Conreign.Core.AI.Battle
             return _map.Select(x => new ReadOnlyPlanetData(x.Value, x.Key)).GetEnumerator();
         }
 
-        public long CalculateDistance(long from, long to)
+        public int CalculateDistance(int from, int to)
         {
             return _map.CalculateDistance(from, to);
         }
