+++ src/Conreign.Core/AI/Battle/ReadOnlyPlanetData.cs
@@ -7,13 +7,13 @@ namespace Conreign.Core.AI.Battle
     {
         private readonly PlanetData _planet;
 
-        public ReadOnlyPlanetData(PlanetData planet, long position)
+        public ReadOnlyPlanetData(PlanetData planet, int position)
         {
             Position = position;
             _planet = planet ?? throw new ArgumentNullException(nameof(planet));
         }
 
-        public long Position { get; }
+        public int Position { get; }
         public string Name => _planet.Name;
         public int ProductionRate => _planet.ProductionRate;
         public double Power => _planet.Power;
